client: Split selfoss-db file
It is nightmare trying to edit all three objects at the same time.

No changes were made other than the split and import wrangling.
